requirements.txt
asgiref==3.4.1
astroid==2.6.5
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.6
django-autoslug==1.9.8
django-cities-light==3.8.1
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==8.3.1
progressbar2==3.53.1
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
